rocket add

+++ dot_editor.js
@@ -1,7 +1,7 @@
+// 컬러픽커 플러그인 설정
 Coloris({
   el: '.coloris',
 });
-
 Coloris.setInstance('.coloris', {
   theme: 'pill',
   themeMode: 'dark',
@@ -11,30 +11,43 @@ Coloris.setInstance('.coloris', {
   swatches: [ '#067bc2', '#84bcda', '#80e377', '#ecc30b', '#f37748', '#d56062' ]
 });
 
+// 도트 컨테이너
 const container = document.getElementById('container');
+// 초기 설정 값
 const setValue = {
   baseWidth: 300,
   baseHeight: 300,
   dotWidth: 10,
   dotHeight: 10
 };
+// 도트 입력 / 지우기 온오프 값
 let dotStatus = true;
 
+// 컨테이너 사이즈 조정을 위한 인풋 선택자
 const setWidth = document.getElementById('setWidth');
 const setHeight = document.getElementById('setHeight');
-
-function setSize() {
-  setValue.baseWidth = setWidth.value;
-  setValue.baseHeight = setHeight.value;
+// 컨ㅔ이너 사이즈 변경 함수
+// function setSize() {
+//   setValue.baseWidth = setWidth.value;
+//   setValue.baseHeight = setHeight.value;
+//   container.style.width = `${setValue.baseWidth}px`;
+//   container.style.height = `${setValue.baseHeight}px`;
+// }
+setWidth.addEventListener('change', function(e) {
+  setValue.baseWidth = e.target.value;
   container.style.width = `${setValue.baseWidth}px`;
+});
+setHeight.addEventListener('change', function(e) {
+  setValue.baseHeight = e.target.value;
   container.style.height = `${setValue.baseHeight}px`;
-}
+});
 
+// coloris 플러그인의 인풋 선택자
 const dotColor = document.getElementById('dotColor');
+// coloris 플러그인의 가변 설정 색상값
 let dotColorVal = dotColor.value;
 
-
-
+// 도트 생성 함수
 function insertDot(x, y, color) {
   const dot = document.createElement('div');
   dot.classList.add('dot');
@@ -43,39 +56,50 @@ function insertDot(x, y, color) {
   dot.style.backgroundColor = color;
   container.append(dot);
 }
-
+// 도트 입력 함수
 function createDot(e) {
+  // 도트가 찍힐 위치값 지정
   const posX = e.offsetX;
   const posY = e.offsetY;
   const dotX = Math.floor(posX / setValue.dotWidth) * setValue.dotWidth;
   const dotY = Math.floor(posY / setValue.dotHeight) * setValue.dotHeight;
   dotColorVal = dotColor.value;
 
+  // 도트 입력 상태일 때
   if (dotStatus === true) {
+    // 클릭 위치가 컨테이너일 때 즉 도트를 찍지 않았으면 도트를 생성
     if (e.target === container) {
       insertDot(dotX, dotY, dotColorVal);
+    // 도트를 찍었다면 해당 도트의 생삭을 변경한다
     } else {
       e.target.style.backgroundColor = dotColorVal;
     }
+  // 도트 지우기 상태일 때
   } else {
+    // 컨테이너를 클릭 하지 않은 상태 즉 도트를 클릭 했을 때만 해당 도트를 삭제
     if (e.target !== container) {
       container.removeChild(e.target);
+    // 도트가 아닌 대상을 클릭 하면 안내문구 출력
     } else {
       console.log('삭제 대상이 없습니다');
     }
   }
 }
+// 컨테이너 클릭할 때 마다 생성된 도트를 입력
 container.addEventListener('click', createDot);
 
+// 도트 지우기 모드로 변경
 function deleteMode() {
   dotStatus = false;
   document.getElementById('status').innerText = 'Delete Mode';
 }
+// 도트 입력 모드로 변경
 function createMode() {
   dotStatus = true;
   document.getElementById('status').innerText = '';
 }
 
+// 마지막에 입력된 도트를 취소함
 function undo() {
   if (container.lastChild){
     container.removeChild(container.lastChild);
@@ -84,13 +108,17 @@ function undo() {
   }
 }
 
+// 도안(배경이미지) 파일 업로드 선택자
 const dotDesign = document.getElementById('dotDesign');
+// 도안 이미지는 변수에 담아서 변경할 수 있다
 let designBackground = '';
 
+// 도안 파일이 업로드 되면 실행
 dotDesign.addEventListener('change', function(e){
   const file = e.target.files[0]
   const fileReader = new FileReader();
   fileReader.readAsDataURL(file);
+  // 이미지 파일이 업로드 되면 배경으로 설정하고 설정값을 입력한다
   fileReader.onload = function(){
     designBackground = `url(${fileReader.result})`;
     container.style.backgroundImage = designBackground;
@@ -98,10 +126,12 @@ dotDesign.addEventListener('change', function(e){
     container.style.backgroundSize = setValue.baseWidth+'px';
     container.style.backgroundRepeat = 'no-repeat';
 
+    // 에디터 오른 쪽에 예시 이미지를 출력한다
     document.getElementById('sampleImg').src = fileReader.result;
   }
 });
 
+// 도안(배경 이미지)을 온오프하는 함수들
 function designOn() {
   container.style.backgroundImage = designBackground;
 }
@@ -109,6 +139,22 @@ function designOff() {
   container.style.backgroundImage = '';
 }
 
+// 도안 위치값/사이즈 수정
+const bgX = document.getElementById('bgX');
+const bgY = document.getElementById('bgY');
+const bgW = document.getElementById('bgW');
+bgX.addEventListener('change', function(e){
+  container.style.backgroundPositionX = e.target.value+'px';
+});
+bgY.addEventListener('change', function(e){
+  container.style.backgroundPositionY = e.target.value+'px';
+});
+bgW.addEventListener('change', function(e){
+  container.style.backgroundSize = e.target.value+'px';
+});
+
+
+// 도안의 컬러값을 출력하기 위해 화면 색상값을 입력받는 함수
 function openEyeDropper() {
   const resultElement = document.getElementById("result");
 
@@ -131,23 +177,32 @@ function openEyeDropper() {
     });
 }
 
+// 한가지 색상으로 모든 도트를 채우는 함수
 function paintFull() {
-  deleteAll();
-  const dotW = setValue.dotWidth;
-  const dotH = setValue.dotHeight;
-  const rowMax = setValue.baseWidth / dotW;
-  const colMax = setValue.baseHeight / dotH;
-  dotColorVal = dotColor.value;
-  for (let i = 0; i < colMax; i++) {
-    for (let j = 0; j < rowMax; j++) {
-      insertDot(dotW*j, dotH*i, dotColorVal);
+  const confirmCheck = confirm('Really?');
+  if (confirmCheck) {
+    deleteAll();
+    const dotW = setValue.dotWidth;
+    const dotH = setValue.dotHeight;
+    const rowMax = setValue.baseWidth / dotW;
+    const colMax = setValue.baseHeight / dotH;
+    dotColorVal = dotColor.value;
+    for (let i = 0; i < colMax; i++) {
+      for (let j = 0; j < rowMax; j++) {
+        insertDot(dotW*j, dotH*i, dotColorVal);
+      }
     }
   }
 }
+// 모든 도트를 지우는 함수
 function deleteAll() {
-  container.innerHTML = '';
+  const confirmCheck = confirm('Really?');
+  if (confirmCheck) {
+    container.innerHTML = '';
+  }
 }
 
+// 도트를 클립보드에 복사하는 함수
 function codeCopy() {
   const codeHTML = `<div class="pixelBox" style="overflow:hidden;position:relative;width:${setValue.baseWidth}px;height:${setValue.baseHeight}px;">${container.innerHTML}</div>`;
   navigator.clipboard.writeText(codeHTML)
